arxiv_search/pyproject.toml
[project]
name = "arxiv-search"
version = "0.1.0"
description = "Generative search experiment on arxiv paper data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "diffusers>=0.35.2",
  "huggingface-hub>=0.36.0",
  "lxml>=6.0.2",
  "numpy>=2.3.4",
  "omegaconf>=2.3.0",
  "polars>=1.34.0",
  "pydantic>=2.12.3",
  "sentence-transformers>=5.1.2",
  "tensorboard>=2.19.0",
  "torch>=2.7.0",
  "torchvision>=0.22.0",
  "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "altair>=5.5.0",
    "jupyter>=1.1.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.1",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
# TODO: Configure selected linter rules if defaults don't work well
# See https://docs.astral.sh/ruff/configuration/

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

